snlptrynew/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
openai
pydantic
pydantic-settings
python-dotenv
orjson
httpx
requests
beautifulsoup4
lxml
newspaper3k
feedparser
pandas
tqdm
python-dateutil
nltk
pytest
yfinance
